Fix Force Logout for all request.
While here removed some testing code

+++ bdpa_in/server/profile/modify_profile_skills.php
@@ -15,6 +15,12 @@ $json_data = file_get_contents('php://input');
 $data = json_decode($json_data);
 
 $payload = $data->payload;
+
+if(force_logout($data->user_id)) {
+    force_logout_response();
+    exit();
+}
+
 $status = modify_profile_skills($payload->user_id, $payload->skills);
 if ($status != "success") {
     echo json_encode(array("error" => $status));
