+++ bdpa_in/server/profile/modify_profile_user_section.php
@@ -16,6 +16,12 @@ $json_data = file_get_contents('php://input');
 $data = json_decode($json_data);
 
 $payload = $data->payload;
+
+if(force_logout($data->user_id)) {
+    force_logout_response();
+    exit();
+}
+
 $status = modify_profile_user_section($payload->user_id, $payload->type, $payload->section);
 if ($status != "success") {
     echo json_encode(array("error" => $status));
